organize zustand folder
